+++ src/templates/watch.js
@@ -1,0 +1,3 @@
+import watch from './watch/watch'
+
+watch()