+++ src/templates/watch/watch.ts
@@ -1,0 +1,10 @@
+export default function() {
+  const wsPath = (<any>window).__marpCliWatchWS
+
+  if (wsPath) {
+    const ws = new WebSocket(wsPath)
+    ws.addEventListener('message', e => {
+      if (e.data === 'reload') location.reload()
+    })
+  }
+}